Add a new way to manipulate registers
This provides a way to read / write registers that makes it more
suitable for scripting.
The method objectify create a tree that represente a device.
We can get a register or a field, which are objects, and use regular
operators to read or update register.
In addition, all properteis defined in SVD file are available.

Signed-off-by: Alexandre Bailon <alexandre.bailon@gmail.com>