Add methods to RegicePeripheral to improve performances
RegiceObject provides mechanism to cache a register.
By default, this is only used for write (i.e user have to call write
to flush data to register).
Read cache is not enabled because we need to have the same registers'
value as the device debugged to be sure that debugging tools are
reliable.
Still, there are some use case (e.g. clock summary) that require a lot
of registers access, which could cause two issues:
- performances, because we read many registers, often many time the same,
  which could be slow over JTAG.
- reliability, because the cpu may run, and update registers in meantime.

This provides two methods, one two configure the cache for peripheral's
registers, and the other one to prefetch peripheral's register content.
Used together, this would speed up many use cases.

Signed-off-by: Alexandre Bailon <alexandre.bailon@gmail.com>