Rework the way to manage cache
RegiceObject provides cache features to speed up performances
when many registers have to be performed.
This was trying to do it smartly but actually this was a mess.
Enabling / disabling cache should be done at higher level
(e.g where the register access are made).
In that way, the one responsible of register access know cache policy
because it is the one which has set it.
This would avoid further mess with caches and make read / write access
reliable.

Signed-off-by: Alexandre Bailon <alexandre.bailon@gmail.com>